chunk of refactor
